requirements.txt
flet>=0.21.0
python-can>=4.3.1
canopen>=2.2.0
matplotlib>=3.7.0
plotly>=5.17.0
numpy>=1.24.0
pandas>=2.0.0
pyserial>=3.5
asyncio-mqtt>=0.16.0
pyyaml>=6.0
